Initial commit
This commit establishes most of the basic features of SoundCheck.  In
particular:

* Tests can be run, using the `check` IO action.

* The `Arbitrary` and `Testable` interfaces are in place and
  functional.  (In particular, functions with `Arbitrary` parameters
  and `Bool` result type are automatically `Testable`.)

* `Arbitrary` implementations exist for `()`, `Bool`, `Nat`, `Int`,
  and `Integer`.

Note that there are two packages, `soundcheck` and `soundcheck-test`.
This is because `idris --testpkg` won't currently compile a viable
test runner on Windows; the `soundcheck` package includes just the
library code (i.e., it's what users should install), while the
`soundcheck-test` package also creates an executable that runs
SoundCheck's self-test suite.