Add `ArbShow` interface
This interface is a weaker version of `Arbitrary` that removes the
need for an `Arbitrary` type to also implement `Show`.  This allows
tests to be parametrized over types for which `Show` implementations
cannot be defined (e.g. function types, or `Type`).

Unfortunately, the general `ArbShow` implementation for types that
implement both `Arbitrary` and `Show` is sufficiently general as to
make it impossible to create a non-overlapping implementation for
specific types.  Until a better solution is found, all types
implementing `ArbShow` must use named implementations and wrap all
tests using those types as parameters in a `using implementation`
block.